Fix D3D conversion warnings
- Direct3D9.cpp: Add explicit casts for uint32_t-to-float (fog values),
  DWORD-to-float (viewport dimensions), float-to-LONG (scissor rect)
- XbPixelShader.cpp: Cast float stage state fields to int for string
  lookup functions, cast uint32_t fog values to float
- Shader.cpp: Use size_t loop variable for container size comparison
- FixedFunctionState.cpp: Cast state to size_t for container comparison

+++ src/core/hle/D3D8/Direct3D9/Shader.cpp
@@ -195,7 +195,7 @@ void ShaderSources::LoadShadersFromDisk() {
 			"// <XBOX SHADER PROGRAM GOES HERE>\n",
 		};
 		int pos = 0;
-		for (int i = 0; i < insertionPoints.size(); i++) {
+		for (size_t i = 0; i < insertionPoints.size(); i++) {
 			auto insertionPoint = insertionPoints[i];
 			auto index = hlsl.find(insertionPoint, pos);
